backend/pyproject.toml
[project]
name = "rfp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.58.0",
    "fastapi[standard]>=0.120.0",
    "ipython>=9.6.0",
    "jupyter>=1.1.1",
    "python-dotenv>=1.1.1",
    "aiofiles>=23.2.1",
    "celery>=5.3.6",
    "redis>=5.0.1",
    "ibm-watsonx-ai>=1.4.2",
    "pymilvus[milvus-lite]>=2.6.2",
    "langchain-ibm>=0.3.19",
    "langchain-core>=0.3.79",
    "langgraph>=1.0.1",
    "langchain>=0.3.27",
    "setuptools<81",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = []
